requirements.txt
--index-url https://pypi.python.org/simple
grpcio-tools
protobuf>=3.8.0
Shapely==1.6.4.post2
pytest
../epl_protobuf/
-e .